+++ src/Exceptions/InvalidListenerException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace HuangYi\Shadowfax\Exceptions;
+
+class InvalidListenerException extends ShadowfaxException
+{
+}